+++ generators/core/templates/.env
