+++ generators/core/templates/.eslintignore
