+++ generators/core/templates/.gitignore
