+++ generators/core/templates/.prettierignore
