+++ generators/core/templates/.vscode/extensions.json
