+++ generators/core/templates/.vscode/settings.json
