+++ generators/core/templates/README.md
